Timeseries V1/requirements.txt
click==8.0.4
colorama==0.4.5
Cython==0.29.33
dataclasses==0.8
Flask==2.0.3
importlib-metadata==4.8.3
itsdangerous==2.0.1
Jinja2==3.0.3
joblib==1.1.1
MarkupSafe==2.0.1
numpy==1.19.5
pandas==1.1.5
patsy==0.5.3
pmdarima==1.8.4
psutil==5.9.4
python-dateutil==2.8.2
pytz==2022.7
PyYAML==6.0
scikit-learn==0.24.2
scipy==1.5.4
six==1.16.0
statsmodels==0.12.2
threadpoolctl==3.1.0
typing_extensions==4.1.1
urllib3==1.26.14
Werkzeug==2.0.3
zipp==3.6.0
